- js for admin; - add meta tags for seo; - save model with seo by transaction; - codereview; - migrations;

+++ common/models/Category.php
@@ -4,6 +4,7 @@ namespace common\models;
 
 use backend\models\User;
 use Yii;
+use yii\caching\TagDependency;
 
 /**
  * This is the model class for table "category".
@@ -17,11 +18,13 @@ use Yii;
  * @property integer $editor_id
  * @property string  $created_at
  * @property string  $updated_at
+ * @property integer $meta_tag_id
  *
  * @property User    $creator
  * @property User    $editor
  * @property Note[]  $notes
  * @property Post[]  $posts
+ * @property MetaTag $meta_tag
  */
 class Category extends \yii\db\ActiveRecord
 {
@@ -30,6 +33,10 @@ class Category extends \yii\db\ActiveRecord
 
 	const CACHE_KEY      = 'modelCategory_';
 	const CACHE_DURATION = 0;
+
+	public $cnt_all;
+	public $cnt_post;
+	public $cnt_note;
 
 	/**
 	 * @inheritdoc
@@ -48,13 +55,16 @@ class Category extends \yii\db\ActiveRecord
 			[['title', 'status', 'alias'], 'required'],
 			[['status', 'creator_id', 'editor_id'], 'integer'],
 			[['created_at', 'updated_at'], 'safe'],
-			[['title', 'alias', 'teaser'], 'string', 'max' => 255],
+			[['title', 'alias'], 'string', 'max' => 255],
+			['teaser', 'string', 'max' => 1000],
 			[
-				['creator_id'], 'exist', 'skipOnError'     => true, 'targetClass' => Admin::className(),
+				['creator_id'], 'exist', 'skipOnError'     => true,
+				                         'targetClass'     => Admin::className(),
 				                         'targetAttribute' => ['creator_id' => 'id'],
 			],
 			[
-				['editor_id'], 'exist', 'skipOnError'     => true, 'targetClass' => Admin::className(),
+				['editor_id'], 'exist', 'skipOnError'     => true,
+				                        'targetClass'     => Admin::className(),
 				                        'targetAttribute' => ['editor_id' => 'id'],
 			],
 		];
@@ -76,6 +86,14 @@ class Category extends \yii\db\ActiveRecord
 			'created_at' => Yii::t('admin', 'Date created'),
 			'updated_at' => Yii::t('admin', 'Date updated'),
 		];
+	}
+
+	/**
+	 * @return \yii\db\ActiveQuery
+	 */
+	public function getMeta_tag()
+	{
+		return $this->hasOne(MetaTag::className(), ['id' => 'meta_tag_id']);
 	}
 
 	/**
@@ -113,6 +131,39 @@ class Category extends \yii\db\ActiveRecord
 	}
 
 	/**
+	 * @param MetaTag $meta_tag
+	 *
+	 * @return bool
+	 * @throws \Exception
+	 * @throws \yii\db\Exception
+	 */
+	public function saveWithMetaKay(MetaTag $meta_tag)
+	{
+		$transaction = self::getDb()->beginTransaction();
+
+		try {
+			self::save();
+
+			if ($this->isNewRecord || !$meta_tag) {
+				$meta_tag = new MetaTag();
+			}
+
+			$meta_tag->entity = self::className();
+			$meta_tag->save();
+
+			$this->meta_tag_id = $meta_tag->getPrimaryKey();
+			self::update();
+
+			$transaction->commit();
+
+			return true;
+		} catch (\Exception $e) {
+			$transaction->rollBack();
+			throw $e;
+		}
+	}
+
+	/**
 	 * @param bool $insert
 	 *
 	 * @return bool
@@ -144,8 +195,7 @@ class Category extends \yii\db\ActiveRecord
 	{
 		parent::afterSave($insert, $changedAttributes);
 
-		$this->clearCacheModel('all');
-		$this->clearCacheModel($this->alias);
+		$this->clearCacheModel();
 	}
 
 	/**
@@ -159,30 +209,85 @@ class Category extends \yii\db\ActiveRecord
 		if (!$data) {
 			$model = self::findAll(['status' => self::STATUS_ACTIVE]);
 
-			/** @var Category $item */
-			foreach ($model as $item) {
-				$data[$item->getPrimaryKey()] = $item;
-			}
+			if ($model) {
+				/** @var Category $item */
+				foreach ($model as $item) {
+					$data[$item->getPrimaryKey()] = $item;
+				}
 
-			Yii::$app->cacheFrontend->set($keyCache, $data, self::CACHE_DURATION);
+				Yii::$app->cacheFrontend->set(
+					$keyCache, $data, self::CACHE_DURATION,
+					new TagDependency(['tags' => self::CACHE_KEY])
+				);
+			}
 		}
 
-		return $data;
+		return $data ? $data : [];
 	}
 
 	/**
-	 * @param string|integer $subKey
-	 *
-	 * @return bool
+	 * @param null|string|integer $subKey
 	 *
 	 * delete all: $subKey = "all"
 	 * delete default: $subKey = "default"
 	 * delete one: $subKey = $model->getPrimaryKey
 	 */
-	public static function clearCacheModel($subKey)
+	public static function clearCacheModel($subKey = null)
 	{
-		$keyCache = self::CACHE_KEY . $subKey;
+		if ($subKey) {
+			$keyCache = self::CACHE_KEY . $subKey;
 
-		return Yii::$app->cacheFrontend->delete($keyCache);
+			Yii::$app->cacheFrontend->delete($keyCache);
+		} else {
+			TagDependency::invalidate(Yii::$app->cache, self::CACHE_KEY);
+		}
+	}
+
+	/**
+	 * @return mixed
+	 */
+	public static function getActiveWithCnt()
+	{
+		$keyCache = self::CACHE_KEY . 'all';
+		$data     = Yii::$app->cacheFrontend->get($keyCache);
+
+		if (!$data) {
+			$model = self::find()
+				->select([
+					'category.id', 'category.title', 'category.alias',
+					'COUNT(post.id) AS cnt_post', 'COUNT(note.id) AS cnt_note',
+				])
+				->leftJoin('post', '`post`.`category_id` = `category`.`id`')
+				->leftJoin('note', '`note`.`category_id` = `category`.`id`')
+				->where([
+					'category.status' => self::STATUS_ACTIVE,
+					'post.status'     => self::STATUS_ACTIVE,
+					'note.status'     => self::STATUS_ACTIVE,
+				])
+				->all();
+
+			/** @var Category $item */
+			foreach ($model as $item) {
+				$data[$item->getPrimaryKey()] = $item;
+			}
+
+			Yii::$app->cacheFrontend->set(
+				$keyCache, $data, self::CACHE_DURATION,
+				new TagDependency(['tags' => self::CACHE_KEY])
+			);
+		}
+
+		return $data;
+
+	}
+
+	public static function getActiveWithCntPost()
+	{
+
+	}
+
+	public static function getActiveWithCntNote()
+	{
+
 	}
 }
